refactor(namespace): introduce Exception/ and Attribute/ subnamespaces
Move exception classes and attribute classes out of the top-level namespace
into dedicated subnamespaces, reducing top-level clutter and surfacing the
distinction between cross-cutting concerns.

Moved:
- Studio\OpenApiContractTesting\InvalidOpenApiSpecException
    -> Studio\OpenApiContractTesting\Exception\InvalidOpenApiSpecException
- Studio\OpenApiContractTesting\InvalidOpenApiSpecReason
    -> Studio\OpenApiContractTesting\Exception\InvalidOpenApiSpecReason
- Studio\OpenApiContractTesting\SpecFileNotFoundException
    -> Studio\OpenApiContractTesting\Exception\SpecFileNotFoundException
- Studio\OpenApiContractTesting\OpenApiSpec
    -> Studio\OpenApiContractTesting\Attribute\OpenApiSpec
- Studio\OpenApiContractTesting\SkipOpenApi
    -> Studio\OpenApiContractTesting\Attribute\SkipOpenApi

InvalidThresholdConfigurationException stays top-level for now (moves to
Coverage/ in a follow-up commit). Production callers that previously relied
on same-namespace resolution to reach the moved classes now carry explicit
use statements.

BREAKING: Pre-1.0 namespace reorg, allowed by CHANGELOG. v0.15.0 target.

PHPUnit 1022/1022 passing. PHPStan level 6 clean. PHP-CS-Fixer clean.
