+++ src/Attribute/OpenApiSpec.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Studio\OpenApiContractTesting;
+namespace Studio\OpenApiContractTesting\Attribute;
 
 use Attribute;
 
 #[Attribute(Attribute::TARGET_CLASS | Attribute::TARGET_METHOD)]
-final class SkipOpenApi
+final class OpenApiSpec
 {
     public function __construct(
-        public readonly string $reason = '',
+        public readonly string $name,
     ) {}
 }
