+++ src/Exception/SpecFileNotFoundException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Studio\OpenApiContractTesting;
+namespace Studio\OpenApiContractTesting\Exception;
 
 use RuntimeException;
 use Throwable;
