+++ tests/Unit/Internal/HttpRefLoaderTest.php
@@ -9,9 +9,9 @@ use GuzzleHttp\Psr7\Response;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\RequestInterface;
+use Studio\OpenApiContractTesting\Exception\InvalidOpenApiSpecException;
+use Studio\OpenApiContractTesting\Exception\InvalidOpenApiSpecReason;
 use Studio\OpenApiContractTesting\Internal\HttpRefLoader;
-use Studio\OpenApiContractTesting\InvalidOpenApiSpecException;
-use Studio\OpenApiContractTesting\InvalidOpenApiSpecReason;
 use Studio\OpenApiContractTesting\Tests\Helpers\FakeHttpClient;
 use Studio\OpenApiContractTesting\Tests\Helpers\FakeHttpClientUnexpectedRequest;
