+++ tests/Unit/Laravel/ExploresOpenApiEndpointTest.php
@@ -7,11 +7,11 @@ namespace Studio\OpenApiContractTesting\Tests\Unit\Laravel;
 use PHPUnit\Framework\AssertionFailedError;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
+use Studio\OpenApiContractTesting\Attribute\OpenApiSpec;
 use Studio\OpenApiContractTesting\Fuzz\ExplorationCases;
 use Studio\OpenApiContractTesting\Fuzz\ExploredCase;
 use Studio\OpenApiContractTesting\HttpMethod;
 use Studio\OpenApiContractTesting\Laravel\ExploresOpenApiEndpoint;
-use Studio\OpenApiContractTesting\OpenApiSpec;
 use Studio\OpenApiContractTesting\OpenApiSpecLoader;
 
 // Re-use the namespace-level config() mock so config()-based fallback works.
