+++ tests/Unit/SkipOpenApiResolverClassLevelTest.php
@@ -6,7 +6,7 @@ namespace Studio\OpenApiContractTesting\Tests\Unit;
 
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
-use Studio\OpenApiContractTesting\SkipOpenApi;
+use Studio\OpenApiContractTesting\Attribute\SkipOpenApi;
 use Studio\OpenApiContractTesting\SkipOpenApiResolver;
 
 #[SkipOpenApi(reason: 'class-level')]
