+++ tests/Unit/SkipOpenApiResolverTest.php
@@ -6,7 +6,7 @@ namespace Studio\OpenApiContractTesting\Tests\Unit;
 
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
-use Studio\OpenApiContractTesting\SkipOpenApi;
+use Studio\OpenApiContractTesting\Attribute\SkipOpenApi;
 use Studio\OpenApiContractTesting\SkipOpenApiResolver;
 
 class SkipOpenApiResolverTest extends TestCase
